requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==3.2.8
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-spectacular==0.22.0
inflection==0.5.1
jsonschema==4.4.0
Pillow==9.1.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
